Renamed repl::parser module to parse.
