+++ core/src/lib.rs
@@ -2,4 +2,5 @@ pub mod datamodel;
 pub mod eval;
 pub mod format;
 pub mod load;
+pub(crate) mod parse;
 pub mod repl;
