+++ core/src/parse/metadata.rs
@@ -1,7 +1,7 @@
 //! Parsers related to metadata aka comments.
 
+use crate::parse::character;
 use crate::repl;
-use repl::parser::character;
 
 use winnow::{
     ascii::{line_ending, space0, space1, till_line_ending},
@@ -135,7 +135,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::repl::parser::testing::expect_parse_ok;
+    use crate::repl::parse::testing::expect_parse_ok;
 
     use pretty_assertions::assert_eq;
